Trying to do a pull-request on myself

+++ src/Savegame/BattleItem.cpp
@@ -1033,9 +1033,9 @@ bool BattleItem::isAmmoVisibleForSlot(int slot) const
  * Get item weight with ammo weight.
  * @return Weight.
  */
-int BattleItem::getTotalWeight() const
+float BattleItem::getTotalWeight() const
 {
-	int weight = _rules->getWeight();
+	float weight = _rules->getWeight();
 	for (const auto* a : _ammoItem)
 	{
 		if (a && a != this)
